requirements.txt
google-api-core==1.31.5
google-cloud-bigquery==2.32.0
google-cloud-storage==1.44.0
apache-airflow==2.3.2
apache-airflow-providers-google==6.5.0
responses==0.20.0
apache-airflow-providers-apache-beam==3.3.0
apache-airflow-providers-sendgrid==2.0.4
croniter==1.3.5